requirements.txt
python-dotenv
hydra-core
datasets
omegaconf
transformers
wandb
nltk
accelerate
matplotlib
seaborn